enable testing
