+++ flexmark-ext-d2-diagrams/src/main/java/com/vladsch/flexmark/ext/d2/D2Extension.java
@@ -3,9 +3,7 @@ package com.vladsch.flexmark.ext.d2;
 import com.vladsch.flexmark.ext.d2.internal.*;
 import com.vladsch.flexmark.html.HtmlRenderer;
 import com.vladsch.flexmark.parser.Parser;
-import com.vladsch.flexmark.util.data.DataKey;
 import com.vladsch.flexmark.util.data.MutableDataHolder;
-import com.vladsch.flexmark.util.data.NullableDataKey;
 
 import org.jetbrains.annotations.NotNull;
