--- flexmark-ext-d2-diagrams/src/test/java/com/vladsch/flexmark/ext/d2/D2SpecTest.java
@@ -1,48 +1,0 @@
-/*
- * Copyright (c) 2015-2016 Vladimir Schneider <vladimir.schneider@gmail.com>, all rights reserved.
- *
- * This code is private property of the copyright holder and cannot be used without
- * having obtained a license or prior written permission of the of the copyright holder.
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- *
- */
-
-package com.vladsch.flexmark.ext.d2;
-
-import com.vladsch.flexmark.core.test.util.RendererSpecTest;
-import com.vladsch.flexmark.parser.Parser;
-import com.vladsch.flexmark.test.util.spec.ResourceLocation;
-import com.vladsch.flexmark.test.util.spec.SpecExample;
-import com.vladsch.flexmark.util.data.DataHolder;
-import com.vladsch.flexmark.util.data.MutableDataSet;
-import org.jetbrains.annotations.NotNull;
-import org.junit.runners.Parameterized;
-
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-public class D2SpecTest extends RendererSpecTest {
-    final private static String SPEC_RESOURCE = "/ext_d2_ast_spec.md";
-    final public static @NotNull ResourceLocation RESOURCE_LOCATION = ResourceLocation.of(D2SpecTest.class, SPEC_RESOURCE);
-    final private static DataHolder OPTIONS = new MutableDataSet()
-            .set(Parser.EXTENSIONS, Collections.singleton(D2Extension.create()))
-            .toImmutable();
-
-    final private static Map<String, DataHolder> optionsMap = new HashMap<>();
-    public D2SpecTest(@NotNull SpecExample example) {
-        super(example, optionsMap, OPTIONS);
-    }
-
-    @Parameterized.Parameters(name = "{0}")
-    public static List<Object[]> data() {
-        return getTestData(RESOURCE_LOCATION);
-    }
-}
