--- flexmark-ext-d2-diagrams/src/test/java/com/vladsch/flexmark/ext/d2/ExtD2TestSuite.java
@@ -1,11 +1,0 @@
-package com.vladsch.flexmark.ext.d2;
-
-import org.junit.runner.RunWith;
-import org.junit.runners.Suite;
-
-@RunWith(Suite.class)
-@Suite.SuiteClasses({
-        D2SpecTest.class,
-})
-public class ExtD2TestSuite {
-}
