+++ flexmark-ext-d2-diagrams/src/test/resources/d2_formatter_spec.md
@@ -1,0 +1,24 @@
+---
+title: D2 Diagrams Extension Formatting Spec
+author: Paweł Biegun
+version: 1.0
+date: '2023.08.26'
+license: '[CC-BY-SA 4.0](http://creativecommons.org/licenses/by-sa/4.0/)'
+...
+
+---
+
+## D2 Diagrams
+
+Converts d2 diagram to svg
+
+```````````````````````````````` example D2 Diagrams: 1
+```d2
+A -> B
+B -> A
+```
+.
+.
+Document[0, 23]
+  D2Block[0, 23]
+````````````````````````````````
